+++ src/com/google/bamboo/soy/annotators/MissingParametersAnnotator.java
@@ -14,7 +14,7 @@
 
 package com.google.bamboo.soy.annotators;
 
-import com.google.bamboo.soy.ParamUtils;
+import com.google.bamboo.soy.lang.ParamUtils;
 import com.google.bamboo.soy.elements.CallStatementBase;
 import com.google.bamboo.soy.parser.SoyTemplateReferenceIdentifier;
 import com.intellij.lang.annotation.AnnotationHolder;
