+++ src/com/google/bamboo/soy/elements/TemplateBlockElement.java
@@ -18,8 +18,8 @@ import com.google.bamboo.soy.parser.SoyAtInjectSingle;
 import com.google.bamboo.soy.parser.SoyAtParamSingle;
 import com.google.bamboo.soy.parser.SoyBeginDelegateTemplate;
 import com.google.bamboo.soy.parser.SoyTemplateDefinitionIdentifier;
-import com.google.bamboo.soy.scope.Scope;
-import com.google.bamboo.soy.scope.Variable;
+import com.google.bamboo.soy.lang.Parameter;
+import com.google.bamboo.soy.lang.Scope;
 import com.google.bamboo.soy.stubs.TemplateBlockStub;
 import com.intellij.psi.PsiNamedElement;
 import com.intellij.psi.StubBasedPsiElement;
@@ -45,5 +45,5 @@ public interface TemplateBlockElement
   boolean isDelegate();
 
   @NotNull
-  List<Variable> getParameters();
+  List<Parameter> getParameters();
 }
