+++ src/com/google/bamboo/soy/lang/ParamUtils.java
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.bamboo.soy;
+package com.google.bamboo.soy.lang;
 
 import com.google.bamboo.soy.elements.CallStatementBase;
 import com.google.bamboo.soy.parser.SoyTemplateBlock;
-import com.google.bamboo.soy.scope.Variable;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.util.PsiTreeUtil;
 import java.util.ArrayList;
@@ -25,13 +24,13 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class ParamUtils {
-  public static List<Variable> getParamDefinitions(PsiElement element) {
+  public static List<Parameter> getParamDefinitions(PsiElement element) {
     SoyTemplateBlock templateBlock = getParentTemplateBlock(element);
     return templateBlock != null ? templateBlock.getParameters() : new ArrayList<>();
   }
 
   /* Only uses stub tree. */
-  public static List<Variable> getParametersForInvocation(PsiElement position, String identifier) {
+  public static List<Parameter> getParametersForInvocation(PsiElement position, String identifier) {
     SoyTemplateBlock templateBlock =
         TemplateNameUtils.findTemplateDeclaration(position, identifier);
 
@@ -54,8 +53,6 @@ public class ParamUtils {
 
   private static SoyTemplateBlock getParentTemplateBlock(PsiElement element) {
     return (SoyTemplateBlock)
-        PsiTreeUtil.findFirstParent(
-            element,
-            psiElement -> psiElement instanceof com.google.bamboo.soy.parser.SoyTemplateBlock);
+        PsiTreeUtil.findFirstParent(element, psiElement -> psiElement instanceof SoyTemplateBlock);
   }
 }
