+++ src/com/google/bamboo/soy/lang/Variable.java
@@ -1,4 +1,4 @@
-package com.google.bamboo.soy.scope;
+package com.google.bamboo.soy.lang;
 
 import com.google.bamboo.soy.parser.SoyParamDefinitionIdentifier;
 import com.google.bamboo.soy.parser.SoyVariableDefinitionIdentifier;
@@ -7,16 +7,14 @@ import com.intellij.psi.PsiNamedElement;
 public class Variable {
   public final String name;
   public final String type;
-  public final boolean isOptional;
   public final PsiNamedElement element;
 
-  public Variable(String name, String type, boolean isOptional, PsiNamedElement element) {
+  public Variable(String name, String type, PsiNamedElement element) {
     assert element instanceof SoyParamDefinitionIdentifier
         || element instanceof SoyVariableDefinitionIdentifier;
 
     this.name = name.replaceFirst("^\\$", "");
     this.type = type;
-    this.isOptional = isOptional;
     this.element = element;
   }
 }
